Add blacklist for float conversion
I now have a topic that publishes mixed-type messages: some are string,
some are float (but for human consumption). This commit provides
a way to prevent the automatic float conversion.

Note that this functionality only applies when a scalar is provided
as measurement, not when a JSON-encoded dict is provided.